+++ tools/run_tests.sh
@@ -67,6 +67,7 @@ classify_test() {
 		*biome_world_generation*|\
 		*forest_isometric_texture_transition*|\
 		*milestone_10_asset_fallback_policy*|\
+		*milestone_11_weapon_drop_progression*|\
 		*milestone_10_cross_biome_chase*|\
 		*milestone_10_full_region_streaming*|\
 		*milestone_10_isometric_performance*|\